chore: refactor make room for more contexts and better border
